Break out the PaginationStrategy into its own file
