+++ lib/brainstem.rb
@@ -6,6 +6,7 @@ require "brainstem/query_strategies/base_strategy"
 require "brainstem/query_strategies/filter_and_search"
 require "brainstem/query_strategies/filter_or_search"
 require "brainstem/query_strategies/paginator"
+require "brainstem/query_strategies/pagination_strategy"
 
 # The Brainstem module itself contains a +default_namespace+ class attribute and a few helpers that make managing +PresenterCollections+ and their corresponding namespaces easier.
 module Brainstem
